requirements.txt
python-telegram-bot==20.3
gigachat==0.1.9
python-dotenv==1.0.0
aiohttp==3.8.1
psycopg2-binary==2.9.6
pillow